+++ stockfish-core/src/core/piece.rs
@@ -89,10 +89,6 @@ impl Piece {
     /// Returns a bitboard containing the squares this piece attacks from the
     /// given `square`, given an `occupancy` bitboard containing all of the
     /// squares with pieces on them that might interfere with its attack.
-    ///
-    /// The `square` the `piece` is on *must not* be in the `occupancy`
-    /// bitboard. If this happens, this function is not guaranteed to produce
-    /// correct results.
     #[inline]
     pub const fn attacks(self, square: Square, board: Bitboard) -> Bitboard {
         accelerate::attacks(